Move preprocess_name out of the director and into the builder.
